+++ packages/simple_graphql_client_gen/CHANGELOG.md
@@ -1,7 +1,11 @@
+## 0.0.3
+
+- fix toString method output
+
 ## 0.0.2
 
-use function type parameter name
+- use function type parameter name
 
 ## 0.0.1
 
-initial release
+- initial release